Create new network package and move files into it.
