+++ google_compute_engine/network/ip_forwarding/ip_forwarding_daemon.py
@@ -35,8 +35,8 @@ from google_compute_engine import constants
 from google_compute_engine import file_utils
 from google_compute_engine import logger
 from google_compute_engine import metadata_watcher
-from google_compute_engine import network_utils
-from google_compute_engine.ip_forwarding import ip_forwarding_utils
+from google_compute_engine.network import network_utils
+from google_compute_engine.network.ip_forwarding import ip_forwarding_utils
 
 LOCKFILE = constants.LOCALSTATEDIR + '/lock/google_ip_forwarding.lock'
